requirements.txt
aniso8601==4.0.1
applescript==2018.11.19
APScheduler==3.5.3
asn1crypto==0.24.0
Authlib==0.10
beautifulsoup4==4.7.1
cachetools==3.0.0
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorclass==2.2.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.5
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
detect==2018.11.19
Django==2.1.3
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
docopt==0.6.2
envparse==0.2.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
fonttools==4.0.2
fpdf==1.7.2
google==2.0.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
gunicorn==19.6.0
html5lib==1.0.1
httpie==1.0.2
httplib2==0.12.0
idna==2.7
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.11.0
Jinja2==2.10
logilab-common==1.4.2
logilab-database==1.15.0
MarkupSafe==1.1.0
names==0.3.0
oauthclient==1.0.3
oauthlib==2.1.0
only==2018.11.20
packaging==19.0
parso==0.1.0
pdfrw==0.4
Pillow==5.3.0
pip-upgrader==1.4.6
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.8
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
ptpython==0.41
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
Pygments==2.2.0
pyparsing==2.3.1
Pyphen==0.9.5
pytesseract==0.2.5
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
restclient==0.11.0
rsa==4.0
runcmd==2018.11.20
six==1.11.0
soupsieve==1.7.3
SQLAlchemy==1.2.14
temp==2018.11.20
terminaltables==3.1.0
tesseract==0.1.3
tinycss2==0.6.1
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
vlc==2018.11.20
wcwidth==0.1.7
WeasyPrint==44
webapp2==2.5.2
webencodings==0.5.1
Werkzeug==0.14.1
yams==0.45.1
Yapps2==2.2.0
zcrmsdk==1.0.6
